修复解析数据的精度问题，解决点击点击算子数据不能跳转的问题

Created-by: insightDeveloper

Match-id-28325739b82cf1e805732052b90aa096567047f7

+++ server/src/modules/memory/parser/MemoryDef.h
@@ -18,7 +18,7 @@ struct MemoryFilePair {
 
 struct Record {
     std::string component;
-    double timesTamp;
+    int64_t timesTamp;
     double totalAllocated;
     double totalReserved;
     std::string deviceType;
@@ -27,8 +27,8 @@ struct Record {
 struct Operator {
     std::string name;
     double size;
-    double allocationTime;
-    double releaseTime;
+    int64_t allocationTime;
+    int64_t releaseTime;
     double duration;
     std::string deviceType;
 };
